+++ src/main/java/ru/vladikshk/myRedis/service/StorageService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface StorageService {
     void put(String key, String value);
-    void put(String key, String value, int expireMs);
+    void put(String key, String value, Long expireMs);
     void remove(String key);
     String get(String key);
     List<String> keys();
